feature: aplicado dry na criacao de usuario e adicionado um handler para exceptions lancadas por conta de keys duplicadas
