+++ src/main/java/SimplifyPay/domain/usecases/DepositMoneyUseCase.java
@@ -1,0 +1,7 @@
+package SimplifyPay.domain.usecases;
+
+public interface DepositMoneyUseCase {
+
+    void execute();
+
+}
